% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darw_circles.R
\name{draw_circles}
\alias{draw_circles}
\title{Generate Points for Drawing Circles}
\usage{
draw_circles(data, x_col = "x", y_col = "y", r_col = "r", n_points = 500)
}
\arguments{
\item{data}{A data frame containing circle data (centers and radii).}

\item{x_col}{Name of the column containing x-coordinates of circle centers.}

\item{y_col}{Name of the column containing y-coordinates of circle centers.}

\item{r_col}{Name of the column containing circle radii.}

\item{n_points}{Number of points to generate around each circle perimeter.}
}
\value{
A dataframe with x, y coordinates for plotting and group identifier per circle plotted.
}
\description{
Creates points along the perimeter of a circle for plotting as a path.
}
